feat: Add order status selection and custom message options
This commit adds the functionality to select order statuses for sending SMS notifications and provides a textarea to enter a custom message. The selected order statuses are stored in the plugin options and checked when an order status is changed. If the new status matches one of the selected statuses, an SMS notification is sent to the customer with the custom message.